Struct constgebra::CMatrix

pub struct CMatrix<const R: usize, const C: usize>(_);

A const matrix type, with dimensions checked at compile time for all operations.

Implementations

impl<const R: usize, const C: usize> CMatrix<R, C>

pub const fn new(vals: [[f64; C]; R]) -> Self

Create a CMatrix from a 2D array of f64.

const ARRAY: [[f64; 2]; 2] = [
    [4.0, 7.0],
    [2.0, 6.0]
];

const CMATRIX: CMatrix::<2, 2> = CMatrix::new(ARRAY);

pub const fn finish(self) -> [[f64; C]; R]

Converts a CMatrix back into a two-dimensional array.

const ARRAY: [[f64; 2]; 2] = [
    [4.0, 7.0],
    [2.0, 6.0]
];

const CMATRIX: CMatrix::<2, 2> = CMatrix::new(ARRAY);

const RESULT: [[f64; 2]; 2] = CMATRIX.finish();

assert_eq!(ARRAY, RESULT)

pub const fn mul<const OC: usize>(self, rhs: CMatrix<C, OC>) -> CMatrix<R, OC>

Multiply two CMatrix and return the result. Columns of self and rows of multiplier must agree in number.

 const LEFT: CMatrix<4, 3> = CMatrix::new([
     [1.0, 0.0, 1.0],
     [2.0, 1.0, 1.0],
     [0.0, 1.0, 1.0],
     [1.0, 1.0, 2.0],
 ]);

 const RIGHT: CMatrix<3, 3> = CMatrix::new([
     [1.0, 2.0, 1.0],
     [2.0, 3.0, 1.0],
     [4.0, 2.0, 2.0]
 ]);

 const EXPECTED: [[f64; 3]; 4] = [
     [5.0, 4.0, 3.0],
     [8.0, 9.0, 5.0],
     [6.0, 5.0, 3.0],
     [11.0, 9.0, 6.0],
 ];

 const RESULT: [[f64; 3]; 4] = LEFT.mul(RIGHT).finish();

 assert_eq!(EXPECTED, RESULT);

pub const fn add(self, rhs: Self) -> Self

Add two CMatrix and return the result.

const LEFT: CMatrix<3, 3> = CMatrix::new([
    [1.0, 0.0, 1.0],
    [2.0, 1.0, 1.0],
    [0.0, 1.0, 1.0]]
);

const RIGHT: CMatrix<3, 3> = CMatrix::new([
    [1.0, 2.0, 1.0],
    [2.0, 3.0, 1.0],
    [4.0, 2.0, 2.0]]
);

const EXPECTED: [[f64; 3]; 3] = [
    [2.0, 2.0, 2.0],
    [4.0, 4.0, 2.0],
    [4.0, 3.0, 3.0]
];

const RESULT: [[f64; 3]; 3] = LEFT.add(RIGHT).finish();

assert_eq!(EXPECTED, RESULT);

pub const fn sub(self, rhs: Self) -> Self

Subtract two CMatrix and return the result.

const LEFT: CMatrix<3, 3> = CMatrix::new([
    [1.0, 2.0, 1.0],
    [2.0, 3.0, 1.0],
    [4.0, 2.0, 2.0]]
);

const RIGHT: CMatrix<3, 3> = CMatrix::new([
    [1.0, 0.0, 1.0],
    [2.0, 1.0, 1.0],
    [0.0, 1.0, 1.0]]
);

const EXPECTED: [[f64; 3]; 3] = [
    [0.0, 2.0, 0.0],
    [0.0, 2.0, 0.0],
    [4.0, 1.0, 1.0]
];

const RESULT: [[f64; 3]; 3] = LEFT.sub(RIGHT).finish();

assert_eq!(EXPECTED, RESULT);

pub const fn apply_each(self, op: Operation) -> Self

Apply an operation to each member of the matrix separately. Especially useful for scaling vectors

const BASE: CMatrix<1, 3> = CMatrix::new([[1.0, 2.0, 3.0]]);
const MUL: CMatrix<1, 3> = BASE.apply_each(Operation::Mul(3.0));
assert_eq!([[3.0, 6.0, 9.0]], MUL.finish());

pub const fn transpose(self) -> CMatrix<C, R>

Return the transpose of a CMatrix.

const START: [[f64; 2]; 2] = [
    [4.0, 7.0],
    [2.0, 6.0]
];

const EXPECTED: [[f64; 2]; 2] = [
    [4.0, 2.0],
    [7.0, 6.0]
];

const RESULT: [[f64; 2]; 2] =
    CMatrix::new(START).transpose().finish();

assert_eq!(EXPECTED, RESULT)

pub const fn givens_l(self, m: usize, a: Sf64, b: Sf64) -> Self

pub const fn givens_r(self, m: usize, a: Sf64, b: Sf64) -> Self

pub const fn svd(self, epsilon: f64) -> CMatrix<C, R>

Singular Value Decomposition

pub const fn pinv(self, epsilon: f64) -> CMatrix<C, R>

impl<const N: usize> CMatrix<1, N>

pub const fn new_vector(vals: [f64; N]) -> CVector<N>

Special case of CMatrix::new for constructing a CVector Always returns a row vector, follow with transpose to build a column vector

const ARRAY: [f64; 2] = [4.0, 7.0];

const ROWVECTOR: CVector::<2> = CVector::new_vector(ARRAY);

pub const fn finish_vector(self) -> [f64; N]

Special case of CMatrix::finish for use with a CVector, returns [f64 ; N] instead of [[f64 ; N]; 1]

const ARRAY: [f64; 2] = [4.0, 7.0];

const CVECTOR: CVector::<2> = CVector::new_vector(ARRAY);

const RESULT: [f64; 2] = CVECTOR.finish_vector();

assert_eq!(ARRAY, RESULT)